Adding doorkeeper

+++ app/controllers/api/products_controller.rb
@@ -1,19 +1,25 @@
 class Api::ProductsController < Api::ApiController
-    # before_action :authenticate_account!
-    # before_action :is_owner?
+    before_action :is_owner?
 
     def index
-        # user=User.find(current_account.accountable_id)
-        # products=user.products
-        products=Product.all
-        render json: products , status: :ok
+        user=User.find(current_account.accountable_id)
+        products=user.products
+        if products
+            render json: products , status: :ok
+        else
+            render json: {message:"No products found"}, status: :not_found
+        end
+        
 
     end
 
     def show
-        product=Product.find(params[:id])
-        render json: product , status: :ok
-
+        product=Product.find_by(id:params[:id])
+        if product
+            render json: product , status: :ok
+        else
+            render json: {message:"product not found"}, status: :not_found
+        end
     end
 
     def new
@@ -26,8 +32,7 @@ class Api::ProductsController < Api::ApiController
         image=productdata["image"]
         productdata.delete("image")
         product=Product.create(productdata)
-                product.user_id=12
-
+        product.user_id=current_account.accountable_id
         product.image.attach(image)
         if product.save
             render json: product ,status: :ok
@@ -37,32 +42,32 @@ class Api::ProductsController < Api::ApiController
     end
 
     def edit
-        product = Product.find(params[:id])
+        product = Product.find_by(id:params[:id])
     end
 
     def update
-        product = Product.find(params[:id])
-        # unless product.user_id==current_account.accountable_id
-        #     redirect_to owner_path
-        #     return
-        # end
+        product = Product.find_by(id:params[:id])
+        unless product.user_id==current_account.accountable_id
+            render json: { message: "You are not authorized for this action"} , status: :forbidden
+            return
+        end
         if product.update(product_params)
             render json: { message: "Updated successfully"}, status: :ok
         else
-             render json: { message: "Error while updating"}, status: :not_found
+            render json: { message: "Error while updating"}, status: :not_found
         end
     end
 
     def destroy
-        product = Product.find(params[:id])
-        # unless product.user_id==current_account.accountable_id
-        #   redirect_to owner_path
-        #   return
-        # end
+        product = Product.find_by(id:params[:id])
+        unless product.user_id==current_account.accountable_id
+            render json: { message: "You are not authorized for this action"} , status: :forbidden
+            return
+        end
         if product.destroy
             render json: { message: "Deleted successfully"}, status: :ok
         else
-             render json: { message: "Error while deleting"}, status: :not_found
+            render json: { message: "Error while deleting"}, status: :not_found
         end
 
     end
@@ -71,15 +76,10 @@ class Api::ProductsController < Api::ApiController
         params.require(:product).permit(:name,:rent,:status,:image,:description)
     end
 
-    # private
-    # def is_owner?
-    #     unless account_signed_in? && current_account.user?
-    #       flash[:alert] = "Unauthorized action"
-    #       if account_signed_in?
-    #           redirect_to renterindex_path
-    #       else
-    #           redirect_to new_account_session_path
-    #       end
-    #     end
-    # end
+    private
+    def is_owner?
+        unless current_account && current_account.user?
+            render json: {message: "You are not authorized to view this page"} , status: :unauthorized
+        end
+    end
 end
